Hotels: Refactor HotelsNavigationOptions
The initial thought with this file was to be reused on other screens.
It turned out this was the only screen needing these options.
There I have moved it to where it belonged and separated
the logic into smaller components
