+++ app/hotels/src/navigation/allHotels/SearchResultsScreen.js
@@ -6,23 +6,18 @@ import {
   type NavigationType,
   useSetNavigationParams,
 } from '@kiwicom/mobile-navigation';
-import {
-  LayoutDoubleColumn,
-  StyleSheet,
-  Icon,
-  Translation,
-} from '@kiwicom/mobile-shared';
+import { LayoutDoubleColumn, StyleSheet } from '@kiwicom/mobile-shared';
 import { defaultTokens } from '@kiwicom/mobile-orbit';
 import { noop } from '@kiwicom/mobile-utils';
 
 import NewAllHotels from '../../allHotels/NewAllHotels';
-import HotelsNavigationOptions from '../HotelsNavigationOptions';
 import {
   SearchResultsContext,
   type SearchResultState,
 } from './SearchResultsContext';
 import SingleHotelContainer from '../../singleHotel/SingleHotelContainer';
 import { HotelsContext, type HotelsContextState } from '../../HotelsContext';
+import NavigationOptions from './navigationOptions/NavigationOptions';
 
 type Props = {|
   +navigation: NavigationType,
@@ -60,49 +55,12 @@ const SearchResultsScreen = (props: Props) => {
   );
 };
 
-SearchResultsScreen.navigationOptions = ({ cityName, navigation }: Props) => {
-  const show = navigation.getParam<string>('show');
-  function goToAllHotelsMap() {
-    const showNext = show === 'list' ? 'map' : 'list';
-    navigation.state.params.toggleMap(showNext);
-  }
-  const icon =
-    show === 'list' ? (
-      <Icon name="map" style={styles.icon} />
-    ) : (
-      <Icon name="list" style={styles.icon} />
-    );
-
-  const translationKey =
-    show === 'list'
-      ? 'hotels_search.all_hotels_search_list.show_map'
-      : 'hotels_search.all_hotels_search_list.show_list';
-  const checkin = navigation.getParam<Date>('checkin');
-  const checkout = navigation.getParam<Date>('checkout');
-  const setCheckinDate =
-    navigation.getParam<() => void>('setCheckinDate') ?? noop;
-  const setCheckoutDate =
-    navigation.getParam<() => void>('setCheckoutDate') ?? noop;
-  return HotelsNavigationOptions({
-    checkin,
-    checkout,
-    cityName,
-    text: <Translation id={translationKey} />,
-    icon,
-    goToAllHotelsMap,
-    setCheckinDate,
-    setCheckoutDate,
-  });
-};
+SearchResultsScreen.navigationOptions = NavigationOptions;
 
 const styles = StyleSheet.create({
   container: {
     flex: 1,
     backgroundColor: defaultTokens.paletteWhite,
-  },
-  icon: {
-    marginEnd: 2,
-    fontSize: 22,
   },
 });
