+++ app/hotels/src/navigation/allHotels/navigationOptions/DateButton.js
