+++ app/hotels/src/navigation/allHotels/navigationOptions/__tests__/DateButton.test.js
